+++ src/connection.c
@@ -33,8 +33,6 @@
 
 static int connection_status = LI_DISCONNECTED;
 
-extern motion_data_state motion_state;
-
 int connection_stage = 0;
 
 bool pause_overlay_is_open(void) {
@@ -99,11 +97,10 @@ static void connection_connection_terminated(int error_code) {
       break;    
   }
 
-  LiStopConnection();
-
   if (connection_status == LI_CONNECTED) {
     stop_output();
   }
+  LiStopConnection();
   vita_debug_log("connection terminated\n");
   connection_status = LI_DISCONNECTED;
   stream_overlay_reset();
@@ -207,32 +204,10 @@ void connection_set_motion_state(uint16_t controller, uint8_t motion_type, uint1
     return;
   }
 
-  if (report_rate == 0) {
-    switch (motion_type) {
-      case LI_MOTION_TYPE_GYRO:
-        motion_state.motion_type_gyro_enabled = false;
-        motion_state.report_rate_gyro = 1;
-        break;
-      case LI_MOTION_TYPE_ACCEL:
-        motion_state.motion_type_accel_enabled = false;
-        motion_state.report_rate_accel = 1;
-        break;
-    }
-  } else {
-
-    switch (motion_type) {
-      case LI_MOTION_TYPE_GYRO:
-        motion_state.motion_type_gyro_enabled = true;
-        motion_state.report_rate_gyro = vita_motion_clamp_report_rate(report_rate);
-        vita_debug_log("Setting gyro state true");
-        break;
-      case LI_MOTION_TYPE_ACCEL:
-        motion_state.motion_type_accel_enabled = true;
-        motion_state.report_rate_accel = vita_motion_clamp_report_rate(report_rate);
-        vita_debug_log("Setting accel state true");
-        break;
-    }
-  }
+  vita_motion_set_state(motion_type, report_rate);
+  vita_debug_log("Motion sensor %u is now %s at %u Hz", motion_type,
+                 report_rate == 0 ? "off" : "on",
+                 report_rate == 0 ? 0 : vita_motion_clamp_report_rate(report_rate));
 }
 
 CONNECTION_LISTENER_CALLBACKS connection_callbacks = {